Fix memory leaks, race conditions, and overflow issues
- Fix GDI font leak in splash window: store HFONT alongside HWND in
  SplashState struct, delete unconditionally in CloseSplash, handle
  label creation failure
- Fix progress window exception safety: wrap TokenExtractor::Extract
  in try/catch to ensure DestroyWindow is always called
- Fix DiscordRpcClient pipe race condition: acquire mutex in Stop()
  before closing pipe, use CancelIo to unblock reader thread
- Fix DiscordRpcClient ReadFrame: validate bytes_read on synchronous
  ReadFile completion to prevent parsing garbage headers
- Fix SendFrame integer truncation: check payload.size() against
  UINT32_MAX before casting
- Fix interval multiplication overflow: use uint64_t intermediate for
  update_interval_s * 1000 in Worker::RunLoop
- Add upper bounds for poll_interval_ms (60s) and update_interval_s
  (3600s) in Config::Load to prevent overflow and API spam

+++ src/worker.cpp
@@ -104,7 +104,7 @@ void Worker::RunLoop() {
 
         // 4. Update presence if changed and enough time has passed
         DWORD now = GetTickCount();
-        DWORD min_interval_ms = config_.update_interval_s * 1000;
+        DWORD min_interval_ms = static_cast<DWORD>(static_cast<uint64_t>(config_.update_interval_s) * 1000);
         bool time_ok;
         bool process_changed;
         {
